REFACTOR: migrate IconButton component test files to typescript
